Fix barrel exports
