+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.0.2] - 2024-11-18
+
+### Fixed
+
+- Fix barrel exports, for proper `index.d.ts` file generation.
+
+
 ## [1.0.1] - 2024-11-18
 
 ### Fixed
