--- src/types/Error.ts
@@ -1,7 +1,0 @@
-type Error = ErrorConstructor & {
-  status: number;
-  message: string;
-  error: string;
-};
-
-export { Error };
