--- src/types/ErrorResponse.ts
@@ -1,5 +1,0 @@
-interface ErrorResponse extends Response {
-  isNativeError: boolean;
-}
-
-export type { ErrorResponse };
