+++ src/types/index.d.ts
@@ -1,0 +1,3 @@
+import type { Native } from '@/types/Native';
+
+export { Native };
