--- src/types/index.ts
@@ -1,8 +1,0 @@
-import { Error } from './Error';
-import { ErrorResponse } from '@/types/ErrorResponse';
-
-declare namespace Native {
-  export { Error, ErrorResponse };
-}
-
-export { Native };
